+++ src/Ephemeris.js
@@ -11,7 +11,12 @@ import {
 } from './utilities/validators'
 
 export default class Ephemeris {
-  constructor({ year=0, month=0, day=0, hours=0, minutes=0, seconds=0, latitude=0.00,  longitude=0.00, height=0.00, key=undefined }={}) {
+  constructor({
+    year=0, month=0, day=0, hours=0, minutes=0, seconds=0,
+    latitude=0.00, longitude=0.00, height=0.00,
+    key=undefined,
+    moonQuarterApproximationValue=1.5
+  }={}) {
     // Assumes UTC time
     // * int year (> 0 C.E.)
     // * int month (0 - 11 || 0 = January, 11 = December)
@@ -25,6 +30,8 @@ export default class Ephemeris {
     // * string OR array[string] key - ex: pass in "venus" or ["mercury", "venus"] or leave blank for all
 
     this._key = validateKey(key)
+
+    this._moonQuarterApproximationValue = moonQuarterApproximationValue
 
     this.Observer = new Observer({latitude: latitude, longitude: longitude, height: height, year: year, month: month, day: day, hours: hours, minutes: minutes, seconds: seconds })
 
@@ -53,7 +60,7 @@ export default class Ephemeris {
       case 'sun':
         return new Sol(body, this.Earth, this.Observer)
       case 'luna':
-        return new Luna(body, this.Earth, this.Observer)
+        return new Luna({body: body, earthBody: this.Earth, observer: this.Observer, quarterApproximationValue: this._moonQuarterApproximationValue})
       case 'heliocentric':
         return new HeliocentricOrbitalBody(body, this.Earth, this.Observer)
       case 'star':
@@ -63,4 +70,36 @@ export default class Ephemeris {
         break
     }
   }
+
+  // static CalculateDailyBody({startYear=0, startMonth=0, startDay=0, endYear=0, endMonth=0, endDay=0, hours=12, minutes=0, latitude=0.00, longitude=0.00, key=null}={}) {
+  //   // Returns an array of ephemerii from startDate to endDate for 12pm noon UTC of each day
+  //   // of the specific body
+  //   const startDate = new Date(Date.UTC(startYear, startMonth, startDay, hours, minutes, 0))
+  //   const endDate = new Date(Date.UTC(endYear, endMonth, endDay, hours, minutes, 0))
+  //   const datesArray = []
+  //
+  //   let currentDate = new Date(startDate)
+  //   while(currentDate <= endDate) {
+  //     datesArray.push(new Date(currentDate))
+  //     currentDate = new Date(currentDate.setDate(currentDate.getDate() + 1))
+  //   }
+  //
+  //   const ephemerisArray = datesArray.map(date => {
+  //     return new Ephemeris(
+  //       {
+  //         year: date.getFullYear(),
+  //         month: date.getMonth(),
+  //         day: date.getDate(),
+  //         hours: date.getHours(),
+  //         minutes: date.getMinutes(),
+  //         seconds: date.getSeconds(),
+  //         latitude: latitude,
+  //         longitude: longitude,
+  //         key: key
+  //       }
+  //     )
+  //   })
+  //
+  //   return ephemerisArray
+  // }
 }