+++ src/utilities/util.js
@@ -1,4 +1,4 @@
-import { RTOH, RTD, RTS, TPI } from '../constants'
+import { RTOH, RTD, RTS, TPI, J2000 } from '../constants'
 
 export const util = {};
 
@@ -355,3 +355,72 @@ util.angles = function (p, q, e, body) {
 
   return body
 };
+
+util.isInModSetRange = ({setPoint=0, halfRangeLength=0, setMin=0, setMax=1, number=0}={}) => {
+  // Given a point within a wrapping set (between setMin and setMax)
+  // and a halfRangeLength to either side of the point
+  // Returns => boolean if the number
+  //            is within wrapped range
+  //            (setPoint - halfRangeLength)...(setPoint + halfRangeLength)
+  //            example:
+  //            set: 0...1
+  //            setPoint: 0.9
+  //            halfRangeLength: 0.2
+  //            number: 0.05
+  //            returns true => 0.05 is between 0.7...0.1 (wrapped)
+  let low = setPoint - halfRangeLength;
+  let high = setPoint + halfRangeLength - setMax;
+  if (low < setMin) {
+    setPoint -= low;
+    number = (number - low) % setMax;
+  } else if (high > setMin) {
+    setPoint -= high;
+    number = (setMax + number - high) % setMax;
+  }
+
+  return (number >= setPoint - halfRangeLength) && (number < setPoint + halfRangeLength)
+}
+
+// Mods even negative numbers
+util.mod = (x, mod) => (x % mod + mod) % mod
+
+util.timeInJulianCenturies = julianDate => {
+  return (julianDate - J2000) / 36525.0
+}
+
+util.decimalDegreesToDMSString = (decimalDegrees) => {
+  // converts decimal degrees to degrees / minutes / seconds
+  // https://stackoverflow.com/a/5786627/6826976
+  // * float decimalDegrees
+  // => returns { degrees, minutes, seconds }
+  let degrees = Math.floor(decimalDegrees)
+  let minfloat = (decimalDegrees - degrees) * 60
+  let minutes = Math.floor(minfloat)
+  let secfloat = (minfloat - minutes) * 60
+  let seconds = Math.round(secfloat)
+
+  // After rounding, the seconds might become 60. These two
+  // if-tests are not necessary if no rounding is done.
+  if (seconds === 60) {
+   minutes++
+   seconds = 0
+  }
+
+  if (minutes === 60) {
+   degrees++
+   minutes = 0
+  }
+
+  return degrees + '\u00B0' +
+         minutes + '\'' +
+         Math.floor(seconds) + '"'
+}
+
+
+util.attachApparentLongitudes = (object, decimalDegrees) => {
+  object.apparentLongitude = decimalDegrees
+  object.apparentLongitudeString = util.decimalDegreesToDMSString(object.apparentLongitude)
+  object.apparentLongitude30String = util.decimalDegreesToDMSString(util.mod(object.apparentLongitude, 30))
+
+  return object
+}