+++ CHANGELOG.md
@@ -1,5 +1,44 @@
 # Changelog
 
+## **v1.1.0**
+
+##### Fixed
+
+##### Added
+
+- Added `observerInstance.Date.dateString` property which returns the formatted date string of the observer
+- Added new moon position properties include phase & shape descriptions and orbit calculations (mean ascending node + mean descending node and mean perigee + mean apogee):
+
+```
+new Ephemeris: {
+  moon: {
+    position: {
+      phaseDecimal: float // property which returns a 0.0 - 0.99 representation of the phase.
+      phaseDaysDistance: float // positive float indicating distance from closest phase quarter
+      shapeDirectionString: string // 'Waxing' or 'Waning' depending on the phaseDecimal value
+      shapeString: string // 'Crescent' or 'Gibbous' depending on the illuminatedFraction
+      quarterApproximationValue: float // the amount (in days) in either direction of a quarter event to allow for determining the approximate quarter phase. Defaults to 1.5 days in either direction of a quarter event.
+      quarterApproximationString: string // 'Entering' or 'Leaving' QUARTER EVENT<'New Moon', 'First Quarter', 'Full Moon', 'Last Quarter'> determined by the phaseDecimal and quarterApproximationValue
+      withinQuarterApproximation: boolean // true/false determined by whether phaseDecimal is inside of quarterApproximationValue
+    }
+    orbit: {
+      meanAscendingNode: {
+        apparentLongitude: float
+        apparentLongitudeString: string
+        apparentLongitude30String: string
+      },
+      meanDescendingNode: {...},
+      meanPerigee: {...},
+      meanApogee: {...}
+    }
+  }
+}
+
+```
+
+
+##### Changed
+
 ## **v1.0.0**
 
 ##### Fixed